move init_model convenience function to __init__.py
