+++ python/snewpy/models/__init__.py
@@ -1,9 +1,69 @@
-
-from . import base, ccsn ,presn
+import logging
 from warnings import warn
+
+from snewpy import get_models, model_path
+from . import ccsn, presn
+
 
 def __getattr__(name):
     if name in dir(ccsn):
         warn(f"{__name__}.{name} is moved to {__name__}.ccsn.{name}", FutureWarning)
         return getattr(ccsn, name)
     raise AttributeError(f"module {__name__} has no attribute {name}")
+
+
+def _init_model(model_name, download=True, download_dir=model_path, **user_param):
+    """Attempts to retrieve instantiated SNEWPY model using model class name and model parameters.
+    If a model name is valid, but is not found and `download`=True, this function will attempt to download the model
+
+    Parameters
+    ----------
+    model_name : str
+        Name of SNEWPY model to import, must exactly match the name of the corresponding model class
+    download : bool
+        Switch for attempting to download model data if the first load attempt failed due to a missing file.
+    download_dir : str
+        Local directory to download model files to.
+    user_param : varies
+        User-requested model parameters used to initialize the model, if one is found.
+        Error checking is performed during model initialization
+
+    Raises
+    ------
+    ValueError
+        If the requested model_name does not match any SNEWPY models
+
+    See Also
+    --------
+    snewpy.models.ccsn
+    snewpy.models.presn
+
+    Example
+    -------
+    >>> from snewpy.models.util import init_model; import astropy.units as u
+    >>> init_model('Nakazato_2013', progenitor_mass=13*u.Msun, metallicity=0.004, revival_time=0*u.s, eos='shen')
+    Nakazato_2013 Model: nakazato-shen-BH-z0.004-s30.0.fits
+    Progenitor mass  : 30.0 solMass
+    EOS              : Shen
+    Metallicity      : 0.004
+    Revival time     : 0.0 ms
+
+    :meta private:
+    """
+    if model_name in dir(ccsn):
+        module = ccsn
+    elif model_name in dir(presn):
+        module = presn
+    else:
+        raise ValueError(f"Unable to find model with name '{model_name}' in snewpy.models.ccsn or snewpy.models.presn")
+
+    try:
+        return getattr(module, model_name)(**user_param)
+    except FileNotFoundError as e:
+        logger = logging.getLogger()
+        logger.warning(f"Unable to find model {model_name} in {download_dir}")
+        if not download:
+            raise e
+        logger.warning(f"Attempting to download model...")
+        get_models(model_name, download_dir)
+        return getattr(module, model_name)(**user_param)
